Moved refactored tests up to /test directory
